fix: refactor project structure and finalized messagebrokermanager
