+++ core/src/main/java/com/zextras/carbonio/files/Files.java
@@ -850,7 +850,7 @@ public final class Files {
   public static final class MessageBroker {
     private MessageBroker() {}
 
-    public static final String USER_STATUS_CHANGED_QUEUE        = "USER_STATUS_CHANGED_QUEUE";
+    public static final String USER_STATUS_CHANGED_QUEUE        = "FILES_USER_STATUS_CHANGED_QUEUE";
 
     public static final class Config {
