+++ core/src/main/java/com/zextras/carbonio/files/config/FilesModule.java
@@ -9,9 +9,6 @@ import com.google.inject.Inject;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
-import com.rabbitmq.client.Channel;
-import com.rabbitmq.client.Connection;
-import com.rabbitmq.client.ConnectionFactory;
 import com.zextras.carbonio.files.cache.CacheHandlerFactory;
 import com.zextras.carbonio.files.dal.repositories.impl.ebean.CollaborationLinkRepositoryEbean;
 import com.zextras.carbonio.files.dal.repositories.impl.ebean.FileVersionRepositoryEbean;
@@ -23,14 +20,11 @@ import com.zextras.carbonio.files.dal.repositories.impl.ebean.UserRepositoryRest
 import com.zextras.carbonio.files.dal.repositories.interfaces.*;
 import com.zextras.carbonio.files.graphql.validators.GenericControllerEvaluatorFactory;
 import com.zextras.carbonio.files.messageBroker.MessageBrokerManagerImpl;
-import com.zextras.carbonio.files.messageBroker.MessageBrokerManager;
+import com.zextras.carbonio.files.messageBroker.interfaces.MessageBrokerManager;
 import com.zextras.filestore.api.Filestore;
 import com.zextras.storages.api.StoragesClient;
 
-import java.io.IOException;
 import java.time.Clock;
-import java.util.Optional;
-import java.util.concurrent.TimeoutException;
 
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
