+++ core/src/main/java/com/zextras/carbonio/files/dal/repositories/impl/ebean/UserRepositoryRest.java
@@ -12,8 +12,7 @@ import com.zextras.carbonio.files.config.FilesConfig;
 import com.zextras.carbonio.files.dal.dao.User;
 import com.zextras.carbonio.files.dal.dao.UserMyself;
 import com.zextras.carbonio.files.dal.repositories.interfaces.UserRepository;
-import com.zextras.carbonio.files.messageBroker.MessageBrokerManager;
-import com.zextras.carbonio.files.messageBroker.entities.UserStatusChangedEvent;
+import com.zextras.carbonio.files.messageBroker.interfaces.MessageBrokerManager;
 import com.zextras.carbonio.usermanagement.UserManagementClient;
 import com.zextras.carbonio.usermanagement.entities.UserId;
 import io.vavr.control.Try;
@@ -27,12 +26,10 @@ public class UserRepositoryRest implements UserRepository {
   private static final Logger logger = LoggerFactory.getLogger(UserRepositoryRest.class);
 
   private final String usermanagementUrl;
-  private final MessageBrokerManager messageBrokerManager;
   private final Cache<User> userCache;
 
   @Inject
-  public UserRepositoryRest(FilesConfig filesConfig, MessageBrokerManager messageBrokerManager, CacheHandler cacheHandler) {
-    this.messageBrokerManager = messageBrokerManager;
+  public UserRepositoryRest(FilesConfig filesConfig, CacheHandler cacheHandler) {
     Properties p = filesConfig.getProperties();
     usermanagementUrl =
         "http://"
