+++ core/src/main/java/com/zextras/carbonio/files/messageBroker/MessageBrokerManagerImpl.java
@@ -4,7 +4,6 @@
 
 package com.zextras.carbonio.files.messageBroker;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.rabbitmq.client.Channel;
@@ -12,8 +11,9 @@ import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
 import com.zextras.carbonio.files.config.FilesConfig;
 import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
-import com.zextras.carbonio.files.messageBroker.consumers.ChangedStatusUserConsumer;
-import com.zextras.carbonio.files.messageBroker.entities.UserStatusChangedEvent;
+import com.zextras.carbonio.files.messageBroker.consumers.ChangedStatusUserFilesConsumer;
+import com.zextras.carbonio.files.messageBroker.interfaces.FilesConsumer;
+import com.zextras.carbonio.files.messageBroker.interfaces.MessageBrokerManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -21,8 +21,19 @@ import java.io.IOException;
 import java.util.Optional;
 import java.util.concurrent.TimeoutException;
 
-import static com.zextras.carbonio.files.Files.MessageBroker.USER_STATUS_CHANGED_QUEUE;
-
+/**
+ * This class is basically a repository for MessageBroker (RabbitMQ).
+ * This class has the responsibility of starting consumers of events that need to be listened to.
+ * The whole MessageBroker integration has been implemented such that listening to events is responsibility of
+ * consumers defined in the project that needs to handle them, while how to push events to queue is defined
+ * inside the project's sdk so external projects will be able to push events in the correct queue without worrying
+ * about project-centric logic.
+ * Note: RabbitMQ doesn't have a built-in check for existing queues and the suggested behaviour is to just create
+ * the queue when necessary since if a queue with the same parameters already exists nothing will change;
+ * since we can't know for sure if the queue already exists and since we can't know for
+ * sure which service will start first/will be up when using said queue, the queue should be created both before
+ * reading from it and before writing to it to eliminate every chance of exceptions from a non-existing queue.
+ */
 @Singleton
 public class MessageBrokerManagerImpl implements MessageBrokerManager {
 
@@ -34,27 +45,23 @@ public class MessageBrokerManagerImpl implements MessageBrokerManager {
 
   @Inject
   public MessageBrokerManagerImpl(FilesConfig filesConfig, NodeRepository nodeRepository){
-    this.connection = createMessageBrokerConnection(filesConfig);
-    this.nodeRepository = nodeRepository;
-  }
-
-  @Override
-  public Connection getConnection() {
-    return connection;
-  }
-
-  private Connection createMessageBrokerConnection(FilesConfig filesConfig){
     ConnectionFactory factory = new ConnectionFactory();
     factory.setHost(filesConfig.getMessageBrokerIp());
     factory.setPort(filesConfig.getMessageBrokerPort());
     factory.setUsername(filesConfig.getMessageBrokerUsername());
     factory.setPassword(filesConfig.getMessageBrokerPassword());
     try {
-      return factory.newConnection();
+      this.connection = factory.newConnection();
     } catch (IOException | TimeoutException e) {
+      // Do not throw in creation but throw on health check and every other method that requires a connection
       logger.error("Can't connect to RabbitMQ", e);
-      return null;
     }
+    this.nodeRepository = nodeRepository;
+  }
+
+  @Override
+  public Connection getConnection() {
+    return connection;
   }
 
   private Channel openMessageBrokerChannel() {
@@ -67,57 +74,33 @@ public class MessageBrokerManagerImpl implements MessageBrokerManager {
     }
   }
 
-  private void createAndStartUserStatusChangedConsumer(Channel channel) throws RuntimeException{
-    // Create consumer
-    ChangedStatusUserConsumer consumer = new ChangedStatusUserConsumer(channel, nodeRepository);
-
+  /**
+   * Creates queue relative to given consumer if it doesn't already exist, then starts consumer to listen on that queue.
+   */
+  private void startConsumer(FilesConsumer filesConsumer) throws RuntimeException{
     try {
-      // Create queue if it doesn't exist
-      // If this queue exists already nothing happens
-      channel.queueDeclare(USER_STATUS_CHANGED_QUEUE, true, false, false, null);
+      // Create queue if it doesn't exist, if this queue exists already nothing happens
+      filesConsumer.getChannel().queueDeclare(filesConsumer.getNameOfQueue(), true, false, false, null);
 
       // Start consumer
-      channel.basicConsume(USER_STATUS_CHANGED_QUEUE, true, consumer);
+      filesConsumer.getChannel().basicConsume(filesConsumer.getNameOfQueue(), true, filesConsumer);
     }catch (IOException e) {
-      logger.error("Can't consume from queue", e);
+      throw new RuntimeException("Can't consume from queue", e);
     }
+  }
+
+  /**
+   * Starts all consumers with a separate channel for each one of them, since if an operation inside handleDelivery of
+   * a consumer is heavy we don't want to block the channel while it executes.
+   * If other consumers are needed just create and start them here.
+   */
+  @Override
+  public void startAllConsumers() {
+    startConsumer(new ChangedStatusUserFilesConsumer(openMessageBrokerChannel(), nodeRepository));
   }
 
   @Override
   public boolean healthCheck() {
     return connection != null && connection.isOpen();
   }
-
-  @Override
-  public void startAllConsumers() {
-    // Open new rabbitMQ channel
-    Channel channel = openMessageBrokerChannel();
-
-    // Try to start consumers
-    createAndStartUserStatusChangedConsumer(channel);
-  }
-
-  // TODO not really used here, only for testing. Remember to remove.
-  @Override
-  public void pushUtil(UserStatusChangedEvent userStatusChangedEvent) {
-    // Open new rabbitMQ channel
-    Channel channel = openMessageBrokerChannel();
-
-    try {
-      // Create queue if it doesn't exist
-      // If this queue exists already nothing happens
-      // Consideration: both reader and writer should create queue if it doesn't exist; this is
-      // because we don't know which service starts first
-      channel.queueDeclare(USER_STATUS_CHANGED_QUEUE, true, false, false, null);
-
-      String message = new ObjectMapper().writeValueAsString(userStatusChangedEvent);
-
-      channel.basicPublish("", USER_STATUS_CHANGED_QUEUE, null, message.getBytes("UTF-8"));
-      channel.close();
-    }catch (IOException | TimeoutException e) {
-      logger.error("Can't push to queue", e);
-    }
-  }
-
-
 }
