+++ core/src/main/java/com/zextras/carbonio/files/messageBroker/interfaces/MessageBrokerFilesEvent.java
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2024 Zextras <https://www.zextras.com>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package com.zextras.carbonio.files.messageBroker.interfaces;
+
+public interface MessageBrokerFilesEvent {
+  public String getNameOfQueue();
+}
