+++ core/src/main/java/com/zextras/carbonio/files/messageBroker/interfaces/MessageBrokerManager.java
@@ -2,15 +2,12 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package com.zextras.carbonio.files.messageBroker;
+package com.zextras.carbonio.files.messageBroker.interfaces;
 
 import com.rabbitmq.client.Connection;
-import com.zextras.carbonio.files.messageBroker.entities.UserStatusChangedEvent;
 
 public interface MessageBrokerManager {
   Connection getConnection();
-
   boolean healthCheck();
   void startAllConsumers();
-  void pushUtil(UserStatusChangedEvent UserStatusChangedEvent);
 }
