+++ core/src/test/java/com/zextras/carbonio/files/messageBroker/MessageBrokerIT.java
@@ -13,7 +13,8 @@ import com.zextras.carbonio.files.api.utilities.DatabasePopulator;
 import com.zextras.carbonio.files.api.utilities.entities.SimplePopulatorTextFile;
 import com.zextras.carbonio.files.dal.dao.ebean.Node;
 import com.zextras.carbonio.files.dal.repositories.interfaces.NodeRepository;
-import com.zextras.carbonio.files.messageBroker.entities.UserStatusChangedEvent;
+import com.zextras.carbonio.files.messageBroker.events.UserStatusChangedEvent;
+import com.zextras.carbonio.files.messageBroker.interfaces.MessageBrokerManager;
 import com.zextras.carbonio.usermanagement.enumerations.UserStatus;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterAll;
